+++ airflow/sensors/web_hdfs_sensor.py
@@ -38,7 +38,7 @@ class WebHdfsSensor(BaseSensorOperator):
         self.webhdfs_conn_id = webhdfs_conn_id
 
     def poke(self, context):
-        from airflow.hooks.webhdfs_hook import WebHDFSHook
+        from airflow.providers.apache.hdfs.hooks.webhdfs import WebHDFSHook
         c = WebHDFSHook(self.webhdfs_conn_id)
         self.log.info('Poking for file %s', self.filepath)
         return c.check_for_path(hdfs_path=self.filepath)
